Added cast to remove warning.
git-svn-id: https://svn.opendtect.org/od/trunk@29156 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/WellAttrib/welltietoseismic.cc
@@ -451,7 +451,7 @@ bool DataPlayer::copyDataToLogSet()
 	const UnitOfMeasure* aiuom = 0;
 	for ( int idx=0; idx<relevantunits.size(); idx++ )
 	{
-	    const float curfactor = relevantunits[idx]->scaler().factor;
+	    const float curfactor = (float) relevantunits[idx]->scaler().factor;
 	    const float eps = curfactor / 100.f;
 	    if ( mIsEqual(curfactor,fact,eps) )
 		aiuom = relevantunits[idx];
